perf(audio): bound browser speed automation
