+++ qcut/apps/web/src/lib/audio/browser-audio-export.ts
@@ -10,189 +10,18 @@ import {
 	clampPlaybackRate,
 	getMediaSourceDuration,
 	getMediaTimelineDuration,
-	mapMediaTimelineTime,
 } from "@/lib/video/video-timing";
 import { calculateAudioPreviewState } from "./audio-preview-state";
+import {
+	buildBrowserAudioAutomation,
+	type BrowserAudioAutomationPoint,
+} from "./browser-audio-automation";
 import {
 	collectBrowserAudioExportClips,
 	decodeBrowserAudioExportClips,
 	type DecodedBrowserAudioExportClip,
 } from "./browser-audio-export-clips";
 import { reverseAudioBuffer } from "./audio-buffer-transform";
-
-const MAX_AUTOMATION_RATE = 20;
-
-interface BrowserAudioAutomationPoint {
-	time: number;
-	outputGain: number;
-	pan: number;
-	denoiseHighpass: number;
-	denoiseLowpass: number;
-	eqLow: number;
-	eqMid: number;
-	eqHigh: number;
-	voiceWarmth: number;
-	voiceClarity: number;
-	voicePresence: number;
-	pitchSemitones: number;
-	formantStrength: number;
-	compressorThreshold: number;
-	compressorRatio: number;
-	compressorMakeup: number;
-	limiterThreshold: number;
-	telephoneDry: number;
-	telephoneWet: number;
-	dryMix: number;
-	reverbMix: number;
-	echoMix: number;
-	playbackRate: number;
-}
-
-function hasAudioKeyframes({
-	settings,
-}: {
-	settings: MediaAudioSettings;
-}): boolean {
-	return Object.values(settings.keyframes ?? {}).some(
-		(keyframes) => (keyframes?.length ?? 0) > 0
-	);
-}
-
-function automationTimes({
-	element,
-	duration,
-	fps,
-}: {
-	element: MediaElement;
-	duration: number;
-	fps: number;
-}): number[] {
-	const sampleRate = Math.min(MAX_AUTOMATION_RATE, Math.max(1, fps));
-	const times = new Set<number>([0, duration]);
-	const settings = element.audio;
-	if (
-		hasAudioKeyframes({
-			settings:
-				settings ??
-				calculateAudioPreviewState({
-					element,
-					timelineTime: element.startTime,
-					fps,
-					duration,
-					masterVolume: 1,
-					muted: false,
-					trackMuted: false,
-					forceMuted: false,
-				}).settings,
-		}) ||
-		(settings?.fadeIn ?? element.audioFadeIn ?? 0) > 0 ||
-		(settings?.fadeOut ?? element.audioFadeOut ?? 0) > 0 ||
-		(element.speedKeyframes?.length ?? 0) > 0 ||
-		clampPlaybackRate(element.playbackRate) !== 1
-	) {
-		const sampleCount = Math.ceil(duration * sampleRate);
-		for (let sample = 1; sample < sampleCount; sample += 1) {
-			times.add(Math.min(duration, sample / sampleRate));
-		}
-	}
-	for (const keyframes of Object.values(settings?.keyframes ?? {})) {
-		for (const keyframe of keyframes ?? []) {
-			times.add(
-				Math.min(duration, Math.max(0, keyframe.frame / Math.max(1, fps)))
-			);
-		}
-	}
-	return [...times].sort((left, right) => left - right);
-}
-
-function pointFromState({
-	element,
-	time,
-	duration,
-	fps,
-}: {
-	element: MediaElement;
-	time: number;
-	duration: number;
-	fps: number;
-}): BrowserAudioAutomationPoint {
-	const state = calculateAudioPreviewState({
-		element,
-		timelineTime: element.startTime + time,
-		fps,
-		duration,
-		masterVolume: 1,
-		muted: false,
-		trackMuted: false,
-		forceMuted: false,
-	});
-	const { settings } = state;
-	const denoiseAmount = settings.denoise.enabled
-		? settings.denoise.amount / 100
-		: 0;
-	const telephoneMix = settings.telephone.enabled
-		? settings.telephone.mix / 100
-		: 0;
-	const reverbMix = settings.reverb.enabled ? settings.reverb.mix / 100 : 0;
-	const echoMix = settings.echo.enabled ? settings.echo.mix / 100 : 0;
-	const pitchEnabled =
-		settings.pitch.enabled && Math.abs(settings.pitch.semitones) >= 0.01;
-	return {
-		time,
-		outputGain: state.outputGain,
-		pan: state.pan,
-		denoiseHighpass: 20 + denoiseAmount * 120,
-		denoiseLowpass: 22_000 - denoiseAmount * 12_000,
-		eqLow: settings.equalizer.enabled ? settings.equalizer.lowGainDb : 0,
-		eqMid: settings.equalizer.enabled ? settings.equalizer.midGainDb : 0,
-		eqHigh: settings.equalizer.enabled ? settings.equalizer.highGainDb : 0,
-		voiceWarmth: settings.voiceEnhance.enabled
-			? settings.voiceEnhance.warmth * 0.08
-			: 0,
-		voiceClarity: settings.voiceEnhance.enabled
-			? settings.voiceEnhance.clarity * 0.08
-			: 0,
-		voicePresence: settings.voiceEnhance.enabled
-			? settings.voiceEnhance.presence * 0.08
-			: 0,
-		pitchSemitones: pitchEnabled ? settings.pitch.semitones : 0,
-		formantStrength: pitchEnabled && settings.pitch.preserveFormants ? 1 : 0,
-		compressorThreshold: settings.compressor.enabled
-			? settings.compressor.thresholdDb
-			: 0,
-		compressorRatio: settings.compressor.enabled
-			? settings.compressor.ratio
-			: 1,
-		compressorMakeup: settings.compressor.enabled
-			? 10 ** (settings.compressor.makeupGainDb / 20)
-			: 1,
-		limiterThreshold: settings.limiter.enabled ? settings.limiter.ceilingDb : 0,
-		telephoneDry: 1 - telephoneMix,
-		telephoneWet: telephoneMix,
-		dryMix: Math.max(0, 1 - Math.max(reverbMix, echoMix)),
-		reverbMix,
-		echoMix,
-		playbackRate: mapMediaTimelineTime({
-			element,
-			localTimelineTime: time,
-			fps,
-		}).playbackRate,
-	};
-}
-
-export function buildBrowserAudioAutomation({
-	element,
-	duration,
-	fps,
-}: {
-	element: MediaElement;
-	duration: number;
-	fps: number;
-}): BrowserAudioAutomationPoint[] {
-	return automationTimes({ element, duration, fps }).map((time) =>
-		pointFromState({ element, time, duration, fps })
-	);
-}
 
 function scheduleParameter({
 	parameter,
